// infodoc-mf_user_managment-7d80db4e84d9/src/app/features/user-list/interfaces/combine-user-info.ts
import { BasicInfoModel } from "../../create/model/basic-info.model";
import { ContactInfoModel } from "../../create/model/contact-info.model";
import { EmployeeInfoModel } from "../../create/model/employee-info.model";
import { UserModel } from "../../../shared/interfaces/user.model";

export interface CombinedUserInfo {
  user?: UserModel | null;
  employee?: EmployeeInfoModel | null;
  person: BasicInfoModel;
  contact?: ContactInfoModel | null;
}